requirements.txt
python-binance==1.0.19
pandas>=2.0.0
ta>=0.11.0
python-dotenv==1.0.1
APScheduler==3.10.4
python-telegram-bot>=22.0
requests>=2.28.0
numpy>=1.24.0
websocket-client>=1.6.0
